Convert files to use datasets versions of penguins and penguins_raw

These functions convert files that use the palmerpenguins package to use the versions of penguins and penguins_raw included in the datasets package in R 4.5.0. They removes calls to library(palmerpenguins) and make necessary changes to some variable names (see Details section below).

Usage

convert_files(input, output, extensions = c("R", "r", "qmd", "rmd", "Rmd"))

convert_files_inplace(input, extensions = c("R", "r", "qmd", "rmd", "Rmd"))

convert_dir(input, output, extensions = c("R", "r", "qmd", "rmd", "Rmd"))

convert_dir_inplace(input, extensions = c("R", "r", "qmd", "rmd", "Rmd"))

Arguments

input

For convert_files() and convert_files_inplace(): A character vector of file paths to convert. For convert_dir() and convert_dir_inplace(): A string with a path to a directory of files to convert.

output

For convert_files(): A character vector of output file paths, or NULL to modify files in place. If provided, must be the same length as input. For convert_dir(): A string with the output directory, or NULL to modify the files in the directory in place.

extensions

A character vector of file extensions to process, defaults to R scripts and R Markdown and Quarto documents.

Value

An invisible list with two components:

changed: A named character vector of output paths for files that were modified, with input paths as the names.

not_changed: A named character vector of output paths for files that were not modified, with input paths as names. Files are not changed if they do not contain references to palmerpenguins (i.e. the patterns) listed in the Details section, or if they do not have one of the specified extensions.

Details

Files are converted by:

Replacing library(palmerpenguins) with ""

Replacing variable names:

bill_length_mm -> bill_len

bill_depth_mm -> bill_dep

flipper_length_mm -> flipper_len

body_mass_g -> body_mass

Replacing ends_with("_mm") with starts_with("flipper_"), starts_with(bill_)

Non-convertible files (those without the specified extensions) are copied to the output location if output is provided, but are not modified.

If the output files or directory do not (yet) exist, they will be created (recursively if necessary).

Replacing ends_with("_mm") with starts_with("flipper_"), starts_with(bill_) ensures that modified R code will always run. starts_with("flipper_") isn't intuitively necessary, as there is only one variable starting with "flipper_", in penguins, but this code will not error inside dplyr::(select), even if flipper_len isn't in the data frame (trying to select flipper_len directly will cause an error if that column isn't in the data frame). In an educational context, we suggest manually editing the converted files to replace starts_with("flipper_") to flipper_len if appropriate. To facilitate this, the functions documented here produce a message indicating the files and line numbers where the ends_with("_mm") substitution was made.

See also

Examples

# Single file
penguin_file <- penguins_examples("penguins.R")
output_file <- withr::local_tempfile(fileext = ".R")
result <- convert_files(penguin_file, output_file)
#> - In /tmp/Rtmpy3g0Qf/file1c29773e94fc.R, ends_with("_mm") replaced on line 7 - please check that the subsitution is appropriate.
#> - Please check the changed output files.
cat(readLines(output_file), sep = "\n") # view changes
#> 
#> library(ggplot2)
#> library(dplyr)
#> 
#> # exploring scatterplots
#> penguins |>
#>   select(body_mass, starts_with("flipper_"), starts_with("bill_")) |>
#>   glimpse()
#> 
#> # Scatterplot example: penguin flipper length versus body mass
#> ggplot(data = penguins, aes(x = flipper_len, y = body_mass)) +
#>   geom_point(aes(color = species, shape = species), size = 2) +
#>   scale_color_manual(values = c("darkorange", "darkorchid", "cyan4"))

# Convert multiple files to new locations
input_files <- c(
  penguins_examples("penguins.R"),
  penguins_examples("analysis/penguins.qmd")
)
output_files <- c(
  withr::local_tempfile(fileext = ".R"),
  withr::local_tempfile(fileext = ".qmd")
)
result <- convert_files(input_files, output_files)
#> - In /tmp/Rtmpy3g0Qf/file1c2976137746.R, ends_with("_mm") replaced on line 7 - please check that the subsitution is appropriate.
#> - Please check the changed output files.
#> - Remember to re-knit or re-render and changed Rmarkdown or Quarto documents.

# Convert all files in a directory
penguins_dir <- penguins_examples_dir()
output_dir <- withr::local_tempdir()
result <- convert_dir(penguins_dir, output_dir)
#> - In /tmp/Rtmpy3g0Qf/file1c29d0a053b/penguins.R, ends_with("_mm") replaced on line 7 - please check that the subsitution is appropriate.
#> - Please check the changed output files.
#> - Remember to re-knit or re-render and changed Rmarkdown or Quarto documents.
result$changed # see which files have changed
#> /home/runner/work/_temp/Library/basepenguins/extdata/analysis/penguins.qmd 
#>                    "/tmp/Rtmpy3g0Qf/file1c29d0a053b/analysis/penguins.qmd" 
#>            /home/runner/work/_temp/Library/basepenguins/extdata/penguins.R 
#>                               "/tmp/Rtmpy3g0Qf/file1c29d0a053b/penguins.R" 

# Modify files in-place
input_file <- penguins_examples("penguins.R")

# Copy file so don't overwrite example provided by package
copy_path <- withr::local_tempfile(fileext = ".R")
file.copy(input_file, copy_path)
#> [1] TRUE
convert_files_inplace(copy_path)
#> - In /tmp/Rtmpy3g0Qf/file1c293e4844e2.R, ends_with("_mm") replaced on line 7 - please check that the subsitution is appropriate.
#> - Please check the changed output files.

if (FALSE)  # don't overwrite the example files provided by the package
  # Get all example files
  input_files <- penguins_examples(recursive = TRUE, full.names = TRUE)

  # Convert them in place
  convert_files_inplace(input_files)
#> - In /home/runner/work/_temp/Library/basepenguins/extdata/penguins.R, ends_with("_mm") replaced on line 7 - please check that the subsitution is appropriate.
#> - Please check the changed output files.
#> - Remember to re-knit or re-render and changed Rmarkdown or Quarto documents.
 # \dontrun{}